docs(integration): add attribution for autologging and gorilla
refactor: move all code from autologging utils to utils so that the attribution is clear and simple.
